+++ src/AppLayouts/index.ts
@@ -1,0 +1,10 @@
+export { default as AppHeader } from './AppHeader';
+export type { AppHeaderProps } from './AppHeader';
+export { default as AppMain } from './AppMain';
+export type { AppMainProps } from './AppMain';
+export { default as AppSidebar } from './AppSidebar';
+export type { AppSidebarProps } from './AppSidebar';
+export { default as ClassicLayout } from './ClassicLayout';
+export type { ClassicLayoutProps } from './ClassicLayout';
+export { default as IslandLayout } from './IslandLayout';
+export type { IslandLayoutProps } from './IslandLayout';